chore: rename function and move tests
